nuggets-18 Kotlin-native API - adding tests and docs for Kotlin
